+++ src/common/ui/radioGroup/radioGroupItem/index.ts
@@ -1,2 +1,1 @@
-export * from './radioGroup'
 export * from './radioGroupItem'
