+++ src/common/ui/radioGroup/radioGroupItem/radioGroupItem.module.scss
@@ -1,0 +1,97 @@
+.container {
+  display: flex;
+  align-items: center;
+}
+
+.item {
+  all: unset;
+
+  position: relative;
+
+  width: 16px;
+  height: 16px;
+
+  background-color: transparent;
+  border: 2px solid var(--color-accent-500);
+  border-radius: 50%;
+
+  &::before {
+    content: '';
+
+    position: absolute;
+    z-index: -1;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+
+    width: 0;
+    height: 0;
+    padding: 0;
+
+    border-radius: 50%;
+  }
+
+  &:hover::before,
+  &:active::before,
+  &:focus-visible::before {
+    width: 36px;
+    height: 36px;
+  }
+
+  &:hover::before {
+    background-color: var(--color-dark-500);
+  }
+
+  &:active::before {
+    background-color: var(--color-accent-900);
+  }
+
+  &:focus-visible::before {
+    background-color: var(--color-dark-700);
+  }
+
+  .indicator {
+    position: relative;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    width: 100%;
+    height: 100%;
+
+    &::after {
+      content: '';
+
+      position: absolute;
+      top: 50%;
+      left: 50%;
+      transform: translate(-50%, -50%);
+
+      display: block;
+
+      width: 11px;
+      height: 11px;
+
+      background-color: var(--color-accent-500);
+      border-radius: 50%;
+    }
+  }
+
+  &[data-disabled] {
+    pointer-events: none;
+    border: 2px solid var(--color-accent-900);
+
+    &::before {
+      content: none;
+    }
+
+    .indicator::after {
+      background-color: var(--color-accent-900);
+    }
+  }
+}
+
+.label {
+  margin-left: 12px;
+}
